Amazon-AWS-EC2-Images - "Describe Images in Region" with advance filtering support - part.3. Signed-off-by: Mihail-Cristian Tusa <mihail-cristian.tusa@hpe.com>
