+++ score-jClouds/src/main/java/io/cloudslang/content/jclouds/entities/CommonFilters.java
@@ -5,7 +5,9 @@ package io.cloudslang.content.jclouds.entities;
  * 6/15/2016.
  */
 public enum CommonFilters {
-    ARCHITECTURE("architecture");
+    ARCHITECTURE("architecture"),
+    HYPERVISOR("hypervisor"),
+    KERNEL_ID("kernel-id");
 
     private String value;
