+++ score-jClouds/src/main/java/io/cloudslang/content/jclouds/entities/InstanceFilters.java
@@ -12,7 +12,6 @@ public enum InstanceFilters {
     GROUP_ID("group-id"),
     GROUP_NAME("group-name"),
     HOST_ID("host-Id"),
-    HYPERVISOR("hypervisor"),
     IAM_INSTANCE_PROFILE_ARN("iam-instance-profile.arn"),
     IMAGE_ID("image-id"),
     INSTANCE_ID("instance-id"),
@@ -23,7 +22,6 @@ public enum InstanceFilters {
     INSTANCE_GROUP_ID("instance.group-id"),
     INSTANCE_GROUP_NAME("instance.group-name"),
     IP_ADDRESS("ip-address"),
-    KERNEL_ID("kernel-id"),
     KEY_NAME("key-name"),
     LAUNCH_INDEX("launch-index"),
     LAUNCH_TIME("launch-time"),
